Add JS Page Manager (pm) for page container
Add a JS IIFE function that loads on the homepage. When page containers
are created, the pm.addPage() function enters a reference to the jQuery
page container element into a pages array, and adds event listeners to
all <a> elements that include the site domain. When a user clicks a
link, the page loads inside the page container element.

When closing the page container, the pm.removePage() function can be
called to remove the jQuery page container object.

Additional checks added to page.php to prevent page containers being
created inside existing page containers.