Add a homepage on theme activation
Create and add a page to WordPress and set this page as a static
homepage on theme activation. Also restrict user options in the admin
Settings - Reading dashboard to prevent the user from making changes.

This ensures the website will always load to a static page and display
the WP 98 desktop.