Add redirection to site domain
When a user enters a specific page (www.domain.com/example/) or queries
(/?cats=true) in the address bar, page windows do not behave as
expected. To fix this, all additions to the address now redirect to
the home page and the required additions are saved temporarily in
cookies. This allows specific pages or 404 errors to be loaded after
the home page has loaded correctly.